Add interval fields to Schedule primary key
This commit includes the migration of adding interval fields to the primary key of the Schedule table in the database. By adding Interval Type and Interval Argument to the primary key, it will allow more flexibility in managing schedules especially those with an occurrence offset. The code has been altered in the migration files and the Schedule model in the PersistenceModels project.